objects/setup: move to trx
