+++ src/tr1/meson.build
@@ -103,7 +103,6 @@ sources = [
   'game/gun/gun_rifle.c',
   'game/inventory.c',
   'game/objects/common.c',
-  'game/objects/setup.c',
   'game/savegame/savegame_bson.c',
   'game/savegame/savegame_legacy.c',
   'game/stats/common.c',
